move classes to finder module
