+++ app/src/main/java/com/github/kitakkun/noteapp/di/AppModule.kt
@@ -2,7 +2,6 @@ package com.github.kitakkun.noteapp.di
 
 import androidx.navigation.NavController
 import com.github.kitakkun.noteapp.ui.page.editor.EditorViewModel
-import com.github.kitakkun.noteapp.ui.page.finder.FinderViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -13,8 +12,5 @@ val appModule = module {
             documentRepository = get(),
             navController = navController,
         )
-    }
-    viewModel {
-        FinderViewModel(get())
     }
 }
