+++ finder/src/main/java/com/github/kitakkun/noteapp/finder/FinderModule.kt
@@ -1,0 +1,8 @@
+package com.github.kitakkun.noteapp.finder
+
+import org.koin.androidx.viewmodel.dsl.viewModelOf
+import org.koin.dsl.module
+
+val finderModule = module {
+    viewModelOf(::FinderViewModel)
+}
