+++ finder/src/main/java/com/github/kitakkun/noteapp/finder/FinderView.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.noteapp.ui.page.finder
+package com.github.kitakkun.noteapp.finder
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -25,6 +25,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.kitakkun.noteapp.customview.preview.PreviewContainer
+import com.github.kitakkun.noteapp.finder.composable.DocumentItem
+import com.github.kitakkun.noteapp.finder.composable.SearchField
+import com.github.kitakkun.noteapp.finder.dialog.ConfirmRemovalDialog
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 @Composable
