+++ finder/src/main/java/com/github/kitakkun/noteapp/finder/FinderViewModel.kt
@@ -1,8 +1,9 @@
-package com.github.kitakkun.noteapp.ui.page.finder
+package com.github.kitakkun.noteapp.finder
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.kitakkun.noteapp.data.repository.DocumentRepository
+import com.github.kitakkun.noteapp.finder.composable.DocumentItemUiState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
